Fix to prevent rain and storm from happening at once
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@1534 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ TileEngine/Tactical Placement GUI.cpp
@@ -68,8 +68,8 @@ UINT32 iTPButtons[ NUM_TP_BUTTONS ];
 extern BOOLEAN gfOverheadMapDirty;
 extern BOOLEAN GetOverheadMouseGridNo( INT16 *psGridNo );
 
-extern int iOffsetHorizontal;
-extern int iOffsetVertical;
+extern UINT16 iOffsetHorizontal;
+extern UINT16 iOffsetVertical;
 
 UINT8	gubDefaultButton = CLEAR_BUTTON;
 BOOLEAN gfTacticalPlacementGUIActive = FALSE;
